Rename Contracts project to Core
- Nuget contains implementations and no contract classes, but domain/core classes
