+++ src/Api.Core/Api.Core.csproj
@@ -1,8 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-  	<TargetFrameworks>netstandard2.0;net5.0</TargetFrameworks>
-    <RootNamespace>Zeiss.PiWeb.Api.Contracts</RootNamespace>
+  	<TargetFrameworks>netstandard2.0;net6.0</TargetFrameworks>
+    <RootNamespace>Zeiss.PiWeb.Api.Core</RootNamespace>
     <TreatWarningsAsErrors>false</TreatWarningsAsErrors>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <PackageVersion>8.0.0</PackageVersion>
@@ -10,12 +10,12 @@
   </PropertyGroup>
 
   <PropertyGroup Label="Assembly information">
-    <AssemblyTitle>Contains contract classes used by the ZEISS PiWeb API.</AssemblyTitle>
-    <AssemblyName>Zeiss.PiWeb.Api.Contracts</AssemblyName>
+    <AssemblyTitle>Contains core classes used by the ZEISS PiWeb API.</AssemblyTitle>
+    <AssemblyName>Zeiss.PiWeb.Api.Core</AssemblyName>
     <AssemblyVersion>8.0.0</AssemblyVersion>
     <Company>Carl Zeiss Industrielle Messtechnik GmbH</Company>
     <Copyright>Copyright © 2020 $(Company)</Copyright>
-    <Description>Contains contract classes used by the ZEISS PiWeb API. The ZEISS PiWeb API nuget depends on this package.</Description>
+    <Description>Contains core classes used by the ZEISS PiWeb API. The ZEISS PiWeb API nuget depends on this package.</Description>
     <FileVersion>$(AssemblyVersion)</FileVersion>
     <InformationalVersion>$(AssemblyVersion)</InformationalVersion>
     <NeutralLanguage>en</NeutralLanguage>
@@ -28,7 +28,7 @@
     <EmbedAllSources>true</EmbedAllSources>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <PackageIconUrl>https://raw.githubusercontent.com/ZEISS-PiWeb/PiWeb-Api/develop/logo128px.png</PackageIconUrl>
-    <PackageId>Zeiss.PiWeb.Api.Contracts</PackageId>
+    <PackageId>Zeiss.PiWeb.Api.Core</PackageId>
     <PackageLicenseFile>LICENSE.txt</PackageLicenseFile>
     <PackageProjectUrl>https://github.com/ZEISS-PiWeb/PiWeb-Api</PackageProjectUrl>
     <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
