+++ src/Api.Rest.Dtos.Tests/Api.Rest.Dtos.Tests.csproj
@@ -25,7 +25,7 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Api.Contracts\Api.Contracts.csproj" />
+        <ProjectReference Include="..\Api.Core\Api.Core.csproj" />
         <ProjectReference Include="..\Api.Rest.Dtos\Api.Rest.Dtos.csproj" />
     </ItemGroup>
