+++ src/Api.Rest.Dtos.Tests/Data/CatalogEntryDtoTest.cs
@@ -16,9 +16,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 	using System.Collections;
 	using System.Globalization;
 	using NUnit.Framework;
-	using Zeiss.PiWeb.Api.Contracts;
 	using Zeiss.PiWeb.Api.Rest.Dtos.Data;
-	using Attribute = Zeiss.PiWeb.Api.Contracts.Attribute;
+	using Attribute = Zeiss.PiWeb.Api.Core.Attribute;
 
 	#endregion
