+++ src/Api.Rest.Dtos.Tests/Data/InspectionPlanSerializationTests.cs
@@ -16,9 +16,9 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests.Data
 	using System.Collections.Generic;
 	using Newtonsoft.Json;
 	using NUnit.Framework;
-	using Zeiss.PiWeb.Api.Contracts;
+	using Zeiss.PiWeb.Api.Core;
 	using Zeiss.PiWeb.Api.Rest.Dtos.Data;
-	using Attribute = Zeiss.PiWeb.Api.Contracts.Attribute;
+	using Attribute = Zeiss.PiWeb.Api.Core.Attribute;
 
 	#endregion
