+++ src/Api.Rest.Dtos/Api.Rest.Dtos.csproj
@@ -59,7 +59,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Api.Contracts\Api.Contracts.csproj" />
+    <ProjectReference Include="..\Api.Core\Api.Core.csproj" />
     <ProjectReference Include="..\Api.Definitions\Api.Definitions.csproj" />
   </ItemGroup>
