+++ src/Api.Rest.Dtos/Data/InspectionPlanDtoBase.cs
@@ -17,10 +17,10 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 	using System.Diagnostics;
 	using System.Text.Json.Serialization;
 	using JetBrains.Annotations;
-	using Zeiss.PiWeb.Api.Contracts;
+	using Zeiss.PiWeb.Api.Core;
 	using Zeiss.PiWeb.Api.Rest.Dtos.Converter;
 	using Zeiss.PiWeb.Api.Rest.Dtos.JsonConverters;
-	using Attribute = Zeiss.PiWeb.Api.Contracts.Attribute;
+	using Attribute = Zeiss.PiWeb.Api.Core.Attribute;
 
 	#endregion
