+++ src/Api.Rest.Dtos/JsonConverters/JsonAttributeArrayConverter.cs
@@ -16,14 +16,14 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.JsonConverters
 	using System.Collections.Generic;
 	using System.Text.Json;
 	using System.Text.Json.Serialization;
-	using Zeiss.PiWeb.Api.Contracts;
+	using Zeiss.PiWeb.Api.Core;
 	using Zeiss.PiWeb.Api.Rest.Dtos.Data;
-	using Attribute = Zeiss.PiWeb.Api.Contracts.Attribute;
+	using Attribute = Zeiss.PiWeb.Api.Core.Attribute;
 
 	#endregion
 
 	/// <summary>
-	/// Specialized <see cref="JsonConverter"/> for <see cref="Contracts.Attribute"/> collections.
+	/// Specialized <see cref="JsonConverter"/> for <see cref="Attribute"/> collections.
 	/// </summary>
 	public sealed class JsonAttributeArrayConverter : JsonConverter<IReadOnlyList<Attribute>>
 	{
