+++ src/Api.Rest.Dtos/JsonConverters/JsonAttributeConverter.cs
@@ -18,15 +18,15 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.JsonConverters
 	using System.Text;
 	using System.Text.Json;
 	using System.Text.Json.Serialization;
-	using Zeiss.PiWeb.Api.Contracts;
+	using Zeiss.PiWeb.Api.Core;
 	using Zeiss.PiWeb.Api.Rest.Dtos.Converter;
 	using Zeiss.PiWeb.Api.Rest.Dtos.Data;
-	using Attribute = Zeiss.PiWeb.Api.Contracts.Attribute;
+	using Attribute = Zeiss.PiWeb.Api.Core.Attribute;
 
 	#endregion
 
 	/// <summary>
-	/// Specialized <see cref="JsonConverter"/> for <see cref="Contracts.Attribute"/>-objects.
+	/// Specialized <see cref="JsonConverter"/> for <see cref="Attribute"/>-objects.
 	/// </summary>
 	public sealed class JsonAttributeConverter : JsonConverter<Attribute>
 	{
