+++ src/Api.Rest/HttpClient/Data/DataServiceRestClient.cs
@@ -19,7 +19,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 	using System.Threading;
 	using System.Threading.Tasks;
 	using JetBrains.Annotations;
-	using Zeiss.PiWeb.Api.Contracts;
+	using Zeiss.PiWeb.Api.Core;
 	using Zeiss.PiWeb.Api.Definitions;
 	using Zeiss.PiWeb.Api.Rest.Common.Client;
 	using Zeiss.PiWeb.Api.Rest.Common.Data;
